+++ src/constants.ts
@@ -1,0 +1,1 @@
+export const IMPORT_HASH_FRONTMATTER_PROPERTY = "rss_hash";